+++ dashboard/src/app/(authenticated)/[tenantId]/(diagram)/components/NodeTypes/DataGroupComponents/Duration.tsx
