+++ dashboard/src/app/(authenticated)/[tenantId]/(diagram)/components/NodeTypes/DataGroupComponents/Entry.tsx
